move decoders and converters from common to mdf_reader
