+++ cdm_reader_mapper/mdf_reader/properties.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 from ..properties import (  # noqa
     numeric_types,
+    numpy_floats,
     numpy_integers,
     object_types,
     pandas_nan_integers,
